+++ lib/src/utils/constants.dart
@@ -6,9 +6,6 @@ class Constants {
   Constants._();
 
   static const kDefaultCameraZoom = 15.0;
-  static const kDefaultSourceName = "Source";
-  static const kDefaultDestinationName = "Destination";
-  static const kDefaultDriverName = "Driver";
   static const kRouteWidth = 5;
   static const kRouteColor = Colors.indigo;
   static const kDefaultMarkerSize = Size.square(150);